// src/common/enums/city-and-province.ts
export enum CityAndProvince {
  Seoul = 'seoul',
  Busan = 'busan',
  Daegu = 'daegu',
  Incheon = 'incheon',
  Gwangju = 'gwangju',
  Daejeon = 'daejeon',
  Ulsan = 'ulsan',
  Sejong = 'sejong',
  Gyeonggi = 'gyeonggi',
  Gangwon = 'gangwon',
  Chungbuk = 'chungbuk',
  Chungnam = 'chungnam',
  Jeonbuk = 'jeonbuk',
  Jeonnam = 'jeonnam',
  Gyeongbuk = 'gyeongbuk',
  Gyeongnam = 'gyeongnam',
  Jeju = 'jeju',
}